feat(sched): scheduler trainer UI v1 + HomeActivity mini-app launcher
HomeActivity becomes MAIN/LAUNCHER and lists the mini-apps (Sort visualizer,
Scheduler trainer, Profiler) as Material 3 cards that route to each one's
dedicated Activity. MainActivity loses MAIN/LAUNCHER but stays available as
the Profiler tile.

SchedActivity is the v1 scheduler UI: pick one of the 7 Maziero algorithms
(FCFS / SJF / RR / SRTF / PRIOc / PRIOp / PRIOd), run, and see the per-tick
Gantt on a Compose Canvas, the final metrics (Tt, Tw, Tr, context switches,
preemptions, total time), and the per-task timing table. The workload + the
per-algorithm quantum are hardcoded to the textbook reference (same 5 tasks
the golden test validates), so what the screen shows is what a professor
grades. A workload editor and a 'predict the next task' game come later.

JNI surface (sched/sched_bridge.cpp + SchedBridge.kt):
  nativeSchedListAlgos()        -> Array<String>
  nativeSchedRunMaziero(idx)    -> JSON { algo, total_time, metrics, tasks[], gantt[] }
Built into the existing libalgoviz.so via a new SCHED_SOURCES group in
CMakeLists.txt. JSON is emitted with the same bench::Json helper the profiler
uses, and parsed on the Kotlin side with org.json (no extra dependency).

Also silences three -Wsign-conversion warnings in sim.h (int -> size_t casts
on sys.tasks[cpu.task] indexing); golden test still passes 7/7.

+++ app/src/main/cpp/sched/sim.h
@@ -265,7 +265,7 @@ inline void phase_schedule(System& sys, StepCtx& ctx) {
     // 3. Termination (detected at the start of the tick after remaining hit 0).
     for (Cpu& cpu : sys.cpus) {
         if (cpu.task == -1) continue;
-        Task& t = sys.tasks[cpu.task];
+        Task& t = sys.tasks[static_cast<std::size_t>(cpu.task)];
         if (t.remaining <= 0) {
             t.state = State::Terminated;
             t.finish = sys.time;
@@ -278,7 +278,7 @@ inline void phase_schedule(System& sys, StepCtx& ctx) {
     // 4. Preemption (per CPU).
     for (Cpu& cpu : sys.cpus) {
         if (cpu.task == -1) continue;
-        Task& cur = sys.tasks[cpu.task];
+        Task& cur = sys.tasks[static_cast<std::size_t>(cpu.task)];
         if (cur.state != State::Running) continue;
 
         bool preempt = false;
@@ -365,7 +365,7 @@ inline void phase_execute(System& sys, StepCtx& ctx) {
     for (size_t c = 0; c < sys.cpus.size(); ++c) {
         Cpu& cpu = sys.cpus[c];
         if (cpu.task != -1) {
-            Task& t = sys.tasks[cpu.task];
+            Task& t = sys.tasks[static_cast<std::size_t>(cpu.task)];
             if (t.state == State::Running) {
                 t.remaining--;
                 if (sys.algo == Algo::Rr) cpu.quantum_left--;
